requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-experimental>=0.3.4
langgraph>=0.2.60
langgraph-checkpoint-sqlite>=2.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
duckduckgo-search<7.0.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
streamlit>=1.40.0
requests>=2.31.0
